refactor(client): break dependencies
