+++ client/v1/module.go
@@ -9,6 +9,8 @@ var (
 	// Module for fx.
 	Module = fx.Options(
 		fx.Provide(grpc.NewClientConn),
+		fx.Provide(grpc.NewTask),
+		fx.Provide(grpc.NewServiceClient),
 		fx.Invoke(grpc.Register),
 	)
 )
