+++ client/v1/transport/grpc/client.go
@@ -1,41 +1,19 @@
 package grpc
 
 import (
-	"bytes"
-	"context"
-
-	"github.com/alexfalkowski/go-service/config"
 	v1 "github.com/alexfalkowski/konfig/api/konfig/v1"
-	"github.com/alexfalkowski/konfig/client"
+	"go.uber.org/fx"
+	"google.golang.org/grpc"
 )
 
-const configFile = "APP_CONFIG_FILE"
+// ServiceClientParams for gRPC.
+type ServiceClientParams struct {
+	fx.In
 
-// Client for v1.
-type Client struct {
-	client v1.ServiceClient
-	cfg    *client.Config
+	Conn *grpc.ClientConn
 }
 
-// Perform getting config.
-func (c *Client) Perform(ctx context.Context) error {
-	req := &v1.GetConfigRequest{
-		Application: c.cfg.Application,
-		Version:     c.cfg.Version,
-		Environment: c.cfg.Environment,
-		Cluster:     c.cfg.Cluster,
-		Command:     c.cfg.Command,
-	}
-
-	resp, err := c.client.GetConfig(ctx, req)
-	if err != nil {
-		return err
-	}
-
-	data, err := config.ReadFileFromEnv(configFile)
-	if err != nil || bytes.Compare(data, resp.Config.Data) != 0 {
-		return config.WriteFileToEnv(configFile, resp.Config.Data)
-	}
-
-	return nil
+// NewServiceClient for gRPC.
+func NewServiceClient(params ServiceClientParams) v1.ServiceClient {
+	return v1.NewServiceClient(params.Conn)
 }
